Load data foundation into BigQuery + fix ward_id dtype
- Cast grievance ward_id to nullable Int64 so the CSV serializes as "14"/""
  instead of "14.0", which BQ INT64 rejected on load.
- gitignore the .varuna virtualenv.
- Loaded into varuna dataset (asia-south1): cities(1), wards(198, GEOGRAPHY,
  100% valid), grievances(766,648, 92.25% ward-matched, 39,940 flood-signal).
- GCP budget: 1000 INR cap with alerts at 50/90/100% (per CLAUDE.md 6).

+++ ingestion/normalize_grievances.py
@@ -130,7 +130,9 @@ def main() -> int:
     # ward join — match on the ~unique raw names only, then map (fast for 700k rows)
     uniq = out["ward_name_raw"].dropna().unique()
     lut = {name: match(name) for name in uniq}
-    out["ward_id"] = out["ward_name_raw"].map(lambda n: lut.get(n, (None,))[0])
+    # nullable integer so the CSV writes "14"/"" (not "14.0") -> loads clean into BQ INT64
+    out["ward_id"] = out["ward_name_raw"].map(
+        lambda n: lut.get(n, (None,))[0]).astype("Int64")
     out["ward_name_canon"] = out["ward_name_raw"].map(
         lambda n: lut.get(n, (None, None))[1])
     match_method = out["ward_name_raw"].map(lambda n: lut.get(n, (None, None, "unmatched"))[2])
